requirements.txt
asgiref==3.2.10
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.0
defusedxml==0.6.0
diff-match-patch==20200713
Django==3.1
django-allauth==0.42.0
django-cors-headers==3.4.0
django-import-export==2.3.0
django-nested-admin==3.3.2
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
djangorestframework==3.11.1
djangorestframework-jwt==1.11.0
drf-yasg==1.17.1
et-xmlfile==1.0.1
idna==2.10
inflection==0.5.0
itypes==1.2.0
jdcal==1.4.1
Jinja2==2.11.2
MarkupPy==1.14
MarkupSafe==1.1.1
mysqlclient==2.0.1
oauthlib==3.1.0
odfpy==1.4.1
openapi-codec==1.3.2
openpyxl==3.0.4
packaging==20.4
psycopg2==2.8.5
pycparser==2.20
PyJWT==1.7.1
pyparsing==2.4.7
python-monkey-business==1.0.0
python3-openid==3.2.0
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
rest-social-auth==4.2.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
simplejson==3.17.2
six==1.15.0
social-auth-app-django==3.4.0
social-auth-core==3.3.3
sqlparse==0.3.1
tablib==2.0.0
uritemplate==3.0.1
urllib3==1.25.10
Werkzeug==1.0.1
xlrd==1.2.0
xlwt==1.3.0
